bridge	R Documentation

Identification of Bridging Nodes in Networks

Description

bridge detects bridging nodes between two or more population groups.

Usage

bridge(x1id, x2id, x1, x2, contact, data, gap...)

Arguments

x1id
identifier of the individual at one end of a link

x2id
identifier of the other individual

x1
the value of the concerned variable of the individual at one end of a link

x2
the value of the concerned variable of the other individual

contact
contact rate between two individuals

data
a data frame, containing the variables in the model

gap
numeric, the individual is identified as a bridging node in the network if the distance between two of his/her/its neighbours is larger than gap

...

Details

The bridge function uses edge list data of contact networks.

Value

The returned value is a table, listing the id, degree, maximum gap among neighbours of the indivduals and whether the individual is a bridging node.

Author(s)

Fei MENG <fei.d.meng@gmail.com>

See Also

summary.bridge provides summary of the output table from bridge function.

Examples

male.id<-c(1, 2, 3, 4, 4, 5, 5, 6, 7)
female.id<-c(8, 9, 10, 10, 11, 12, 10, 13, 14)
male.age <- c(15, 20, 18, 40, 40, 26, 26, 25, 23)
female.age <- c(19, 14, 15, 15, 45, 24, 15, 25, 22)
contact<-c(1, 1.2, 0.5, 0.2, 0.8, 1.4, 1, 2, 1.1)
r<-data.frame(male.id,female.id, male.age,female.age,contact)
gap<-10
this.bridge<-bridge(male.id,female.id,male.age,female.age,contact,data=r,gap=10)
print(this.bridge)