contact.beta	R Documentation

Contact Rate Matrix

Description

contact.beta constructs a contact rate matrix based on edge list data

Usage

contact.beta(x1, x2, contact, cate, data, directed = TRUE, ...)

Arguments

x1
the value of the concerned variable of the individual at one end of a link

x2
the value of the concerned variable of the other individual

contact
contact rate between two individuals

data
a data frame, containing the variables in the model

cate
a vector, specifying the cutpoints of subgroup division

directed
logical, indicating whether the network is directed

...

Value

The returned value is a matrix.