matrix2list	R Documentation

Matrix to Edge List Convertion

Description

matrix2list converts contact matrix to an edge list

Usage

matrix2list(m,directed=TRUE,head=TRUE,...)

Arguments

m
the matrix to be converted

directed
logical, indicating whether the contacts in the network is symmetric

head
logical, indicating whether the first row/column is the identifier of the nodes

...

Value

This function returns a three-column data frame (see examples).

Author(s)

Fei MENG <fei.d.meng@gmail.com>

Examples

# m is a contact matrix
m <- matrix(c(0, 1, 0, 4, 0, 0, 0, 0, 1, 2, 1, 0),nrow=3)
rownames(m)<-c('male.1','male.2','male.3')
colnames(m)<-c('female.1','female.2','female.3','female.4')

# the heterosexual contact network is not symmetric
# the first row and the first column are not 
# identifiers of the individuals
edge.list<-matrix2list(m,directed=TRUE,head=FALSE)

# edge.list
# x1.id    x2.id    contact
# male.2 female.1       1
# male.1 female.2       4
# male.3 female.3       1
# male.1 female.4       2
# male.2 female.4       1