neighbour	R Documentation

Neighbourhood of a Set of Nodes

Description

neighbourhood looks for the nodes (neighbours) that are linked to a given set of nodes (origins).

Usage

neighbour(x1.id,x2.id, root.nodes, data, size)

Arguments

x1.id
name of the original nodes

x2.id
name of the neighbouring nodes

root.node
a vector, specifying the origin nodes

data
a data.frame, containing the edge list of a network

size
the maximum size of the returned neighbourhood including the origins. If the value of size is smaller than the neighbourhood, a subset of neighbours will be randomly selected.

Value

a data frame, containing the neighbourhood

Author(s)

Fei MENG <fei.d.meng@gmail.com>

See Also

sampling,sampling.method

Examples

#