sampling	R Documentation

Network Sampling

Description

sampling samples subgraphs from networks.

Usage

sampling(id.names,data,method)

Arguments

id.names
a vector of strings, specifying the identifier of nodes

data
a data frame, containing the edge list of a network

method
a vector, containing the name and parameters of an existing network sampling method

Value

The returned value is a sampled subset of the original data frame.

Author(s)

Fei MENG <fei.d.meng@gmail.com>

See Also

sampling.method