sampling.method	R Documentation

Network Sampling Method

Description

sampling.method examine whether the input vector contains name and parameters of a valid sampling method

Usage

sampling.method(method)

Arguments

method
a vector, whose first elements are the name and paramteters of a sampling method

Details

Currently implemented methods are 'induced','incidenced','star','snowball' and 'link tracing'. Valid formats of method vectors are: method = c('induced', p.node); method = c('incident', p.edge); method = c('star', start.size); method = c('snowball', size, start.size,increase,resample); method = c('tracing', size, start.size);

Value

The returned value is a logical number.

Author(s)

Fei MENG <fei.d.meng@gmail.com>

See Also

sampling, neighbour