sd.jackknife	R Documentation

Standard Error Using Jackknife Method

Description

sd.jackknife estimates the standard error of a network mixing coefficient using a jackknife method.

Usage

sd.jackknife(x1, x2, contact, data, cate, method, ...)

Arguments

x1
the value of the concerned variable of the individual at one end of a link

x2
the value of the concerned variable of the other individual

contact
contacts between two individuals

data
a data frame, containing the variables in the model

cate
a vector, specifying the cutpoints of subgroup division

method
a string, the name of the function that calculate mixing coefficient. Available choices are 'assort.Newman','assort.Gupta','assort.Farrington'

...

Value

The returned value is numeric.

Author(s)

Fei MENG <fei.d.meng@gmail.com>

See Also

assort.Newman, assort.Gupta, assort.Farrington, sd.kappa