sd.kappa	R Documentation

Standard Error of Newman's Coefficient using Cohen's Kappa Method

Description

sd.kappa estimates the standard error of Newman's mixing coefficient using a formula that is adapted from Cohen's Kappa.

Usage

sd.kappa(x1, x2, contact, data, cate, method, ...)

Arguments

x1
the value of the concerned variable of the individual at one end of a link

x2
the value of the concerned variable of the other individual

contact
contacts between two individuals

data
a data frame, containing the variables in the model

cate
a vector, specifying the cutpoints of subgroup division

...

Value

The returned value is numeric.

Author(s)

Fei MENG <fei.d.meng@gmail.com>

References

Cohen, Jacob (1960). "A coefficient of agreement for nominal scales". Educational and Psychological Measurement 20 (1): 37–46. doi:10.1177/001316446002000104

See Also

assort.Newman